requirements.txt
typer==0.12.3
click==8.1.7
pydantic==2.9.2
PyYAML==6.0.2
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
google-api-python-client==2.136.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
tenacity==8.5.0
playwright==1.47.0
python-dotenv==1.0.1
pytest==8.3.2
aiofiles==24.1.0
rich==13.8.1
Unidecode==1.3.8
